Fix test file format (#152)
* rename file & add readme

* update some comment and add a try except

* fix bug in url

* fix bug in api_test

---------
