+++ README.md
@@ -16,6 +16,7 @@ English | [简体中文](./README_zh-CN.md)
 - [How to use](#how-to-use)
     - [1. Model Training](#1-model-training)
     - [2. Inference](#2-inference)
+- [API test](./api_test/README.md)
 - [Algorithm Detailed](#algorithm-detailed)
     - [1. Architectural Overview](#1-architectural-overview)
     - [2. Training Detailed](#2-training-detailed)
