+++ README_zh-CN.md
@@ -16,6 +16,7 @@
 - [如何使用](#如何使用)
     - [1. 模型训练](#1-模型训练)
     - [2. 人物生成](#2-人物生成)
+- [API测试](./api_test/README.md)
 - [算法详细信息](#算法详细信息)
     - [1. 架构概述](#1-架构概述)
     - [2. 训练细节](#2-训练细节)
